Use requestIdleCallback with requestAnimationFrame fback
